% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/days_active_data.R
\name{days_active_data}
\alias{days_active_data}
\title{Count number of active days for each user}
\usage{
days_active_data(
  uuid_data = get_rapidpro_uuid_names(),
  flow_name,
  rapidpro_site = get_rapidpro_site(),
  token = get_rapidpro_key(),
  flatten = FALSE,
  read_runs = FALSE,
  runs_data = "result_flow_runs.RDS",
  include_archived_data = FALSE,
  get_by = "gotit",
  data_from_archived = archived_data,
  read_archived_data_from = "archived_data_monthly.RDS"
)
}
\arguments{
\item{uuid_data}{A string containing UUID values. See \code{set_rapidpro_uuid_names()} to set this value.}

\item{flow_name}{A string containing flow names to call data from.}

\item{rapidpro_site}{A string containing the rapidpro website to call the data from. See \code{set_rapidpro_site()} to amend the website.}

\item{token}{A string containing the token to call the data. See \code{set_rapidpro_key()} to amend the token.}

\item{flatten}{Default \code{FALSE}. A boolean denoting whether the data should be flattened into a two-dimensional tabular structure.}

\item{read_runs}{Default \code{FALSE}. A boolean denoting whether to read in the \code{runs_data}, or to read the data from RapidPro}

\item{runs_data}{If the "runs.json" data has been read in already, this is the run data. Instead, the \code{uuid_data}, \code{call_type}, and \code{token} can be given (note that this latter approach is slower since it requires reading the data in).}

\item{include_archived_data}{Default \code{FALSE}. A boolean denoting whether to include archived data or not.}

\item{get_by}{A string denoting how to access the archived data (if \code{include_archived_data} is \code{TRUE}). Options are \code{"read_from_RDS"}, \code{"gotit"}. The first denotes that the data is read from an RDS file (given in \code{read_archived_data_from}), the latter that the data is stored and will be read in directly (see \code{data_from_archived}).}

\item{data_from_archived}{If \code{get_by = "gotit"}, this is the archived data.}

\item{read_archived_data_from}{If \code{get_by = "read_from_RDS"}, this is the archived data.}
}
\value{
A data frame with two columns: The unique ID of the individual, and the number of days they have been active.
}
\description{
The number of days that each user is active
}
