% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_flows.R
\name{filter_date_range}
\alias{filter_date_range}
\title{Helper function to filter data by date range.}
\usage{
filter_date_range(data, date, format, tzone, start = TRUE)
}
\arguments{
\item{data}{A data frame to be filtered.}

\item{date}{The date to filter the data by.}

\item{format}{The date-time format as used by \code{strptime}.}

\item{tzone}{The time zone specification to be used for conversion.}

\item{start}{A logical value indicating whether to filter for dates before or after the specified date.}
}
\value{
A filtered data frame.
}
\description{
This function filters a data frame based on a specified date range.
}
