% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_data_calculation.R
\name{flow_data_calculation}
\alias{flow_data_calculation}
\title{Calculation to get the flow data.}
\usage{
flow_data_calculation(
  result_flow,
  flatten = FALSE,
  flow_type = "none",
  date_from = NULL,
  date_to = NULL,
  format_date = "\%Y-\%m-\%d",
  tzone_date = "UTC",
  created_on = FALSE
)
}
\arguments{
\item{result_flow}{Data frame containing flow data.}

\item{flatten}{Default \code{FALSE}. A boolean denoting whether the data should be flattened into a two-dimensional tabular structure.}

\item{flow_type}{Default \code{none}. Takes values \code{"praise"}, \code{"calm"}, \code{"check_in"}, \code{"tips"}, \code{"none"}. These are related to ParentText analysis.}

\item{date_from}{character string giving the date to filter the data from.}

\item{date_to}{character string giving the date to filter the data to.}

\item{format_date}{from \verb{as.POSIX*} function: character string giving a date-time format as used by \code{strptime}.}

\item{tzone_date}{from \verb{as.POSIX*} function: time zone specification to be used for the conversion, if one is required. System-specific (see time zones), but "" is the current time zone, and "GMT" is UTC (Universal Time, Coordinated). Invalid values are most commonly treated as UTC, on some platforms with a warning.}

\item{created_on}{Parameter for flow_data.}
}
\value{
Data frame displaying the flow data.
}
\description{
This is called in the \code{get_flow_data} function. This function arranges the function for different flow types.
}
