% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_archived_data.R
\name{get_archived_data}
\alias{get_archived_data}
\title{Get archived (flow/run) data}
\usage{
get_archived_data(
  rapidpro_site = get_rapidpro_site(),
  token = get_rapidpro_key(),
  period = "monthly",
  flatten = FALSE,
  date_from = NULL,
  date_to = NULL,
  archive_type = c("run", "messages"),
  format_date = "\%Y-\%m-\%d",
  tzone_date = "UTC"
)
}
\arguments{
\item{rapidpro_site}{A string containing the rapidpro website to call the data from. See \code{set_rapidpro_site()} to amend the website.}

\item{token}{A string containing the token to call the data. See \code{set_rapidpro_key()} to amend the token.}

\item{period}{Default \code{"monthly"}. A character string giving the period of time. One of \code{"monthly"}, \code{"daily"}, or \code{"none"}.}

\item{flatten}{Default \code{FALSE}. A boolean denoting whether the data should be flattened into a two-dimensional tabular structure.}

\item{date_from}{character string giving the date to filter the data from.}

\item{date_to}{character string giving the date to filter the data to.}

\item{archive_type}{character string denoting whether to call \code{"runs"} or \code{"messages"} data.}

\item{format_date}{from \verb{as.POSIX*} function: character string giving a date-time format as used by \code{strptime}.}

\item{tzone_date}{from \verb{as.POSIX*} function: time zone specification to be used for the conversion, if one is required. System-specific (see time zones), but "" is the current time zone, and "GMT" is UTC (Universal Time, Coordinated). Invalid values are most commonly treated as UTC, on some platforms with a warning.}
}
\value{
Archived data
}
\description{
Download the archived run data from RapidPro.
}
