% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_flow_data.R
\name{get_flow_data}
\alias{get_flow_data}
\title{Get data from each run (flow)}
\usage{
get_flow_data(
  uuid_data = get_rapidpro_uuid_names(),
  flow_name,
  call_type = "runs.json?flow=",
  rapidpro_site = get_rapidpro_site(),
  token = get_rapidpro_key(),
  flatten = FALSE,
  checks = FALSE
)
}
\arguments{
\item{uuid_data}{A string containing UUID values. See \code{set_rapidpro_uuid_names()} to set this value.}

\item{flow_name}{A string containing flow names to call data from.}

\item{call_type}{A string containing the call type.}

\item{rapidpro_site}{A string containing the rapidpro website to call the data from. See \code{set_rapidpro_site()} to amend the website.}

\item{token}{A string containing the token to call the data. See \code{set_rapidpro_key()} to amend the token.}

\item{flatten}{Default \code{FALSE}. A boolean denoting whether the data should be flattened into a two-dimensional tabular structure.}

\item{checks}{Default \code{FALSE}. A boolean whether to check and update the data if it is not found.}
}
\value{
List separated by each flow_name provided. Each element in the list contains a data frame for each flow_name provided.
}
\description{
Call RapidPro data from each run or flow.
}
