% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rapidpro_key.R
\name{get_rapidpro_key}
\alias{get_rapidpro_key}
\title{Get the key (token) from the environment}
\usage{
get_rapidpro_key()
}
\value{
returns a string containing the key (Token) set in the environment for the rapidPro data.
}
\description{
Call the key (token) from the environment. The key is set in the function \code{set_rapidpro_key}.
}
