% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rapidpro_uuid.R,
%   R/get_rapidpro_uuid_names.R
\name{get_rapidpro_uuid}
\alias{get_rapidpro_uuid}
\title{Get the UUID names from the environment}
\usage{
get_rapidpro_uuid()

get_rapidpro_uuid()
}
\value{
returns the UUID names set in the environment for the RapidPro data.

returns the UUID names set in the environment for the RapidPro data.
}
\description{
Call the UUID names from the environment. This set in the function \code{set_rapidpro_uuid_names}.

Call the UUID names from the environment. This set in the function \code{set_rapidpro_uuid_names}.
}
