% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rapidpro_uuid_names.R
\name{get_rapidpro_uuid_names}
\alias{get_rapidpro_uuid_names}
\title{Get the UUID names from the environment}
\usage{
get_rapidpro_uuid_names()
}
\value{
returns the UUID names set in the environment for the rapidPro data
}
\description{
Get the UUID names from the environment
}
