% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_survey_data.R
\name{get_survey_data}
\alias{get_survey_data}
\title{Get the survey data}
\usage{
get_survey_data(parenting_variable, survey_max = 9)
}
\arguments{
\item{parenting_variable}{A character vector containing ...}

\item{survey_max}{A numerical value denoting ...}
}
\value{
A list of survey data
}
\description{
Get the survey data
}
