% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_user_group_data.R
\name{get_user_group_data}
\alias{get_user_group_data}
\title{Get data detailing the user group information.
Is this function needed? We do not use it anymore.}
\usage{
get_user_group_data(user_data = get_user_data(), name = NULL, uuid = NULL)
}
\arguments{
\item{user_data}{}

\item{name}{}

\item{uuid}{}
}
\value{

}
\description{
Get data detailing the user group information.
Is this function needed? We do not use it anymore.
}
