% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_flows.R
\name{handle_calm_flow}
\alias{handle_calm_flow}
\title{Helper function to handle calm flow data.}
\usage{
handle_calm_flow(data)
}
\arguments{
\item{data}{A data frame containing calm flow data.}
}
\value{
A tibble with relevant information.
}
\description{
This function processes calm flow data and extracts relevant information.
}
