% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_flows.R
\name{handle_check_in_flow}
\alias{handle_check_in_flow}
\title{Helper function to handle check-in flow data.}
\usage{
handle_check_in_flow(data)
}
\arguments{
\item{data}{A data frame containing check-in flow data.}
}
\value{
A tibble with relevant information.
}
\description{
This function processes check-in flow data and extracts relevant information.
}
