% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_flows.R
\name{handle_praise_flow}
\alias{handle_praise_flow}
\title{Helper function to handle praise flow data.}
\usage{
handle_praise_flow(data)
}
\arguments{
\item{data}{A data frame containing praise flow data.}
}
\value{
A tibble with relevant information.
}
\description{
This function processes praise flow data and extracts relevant information.
}
