% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_flows.R
\name{handle_tips_flow}
\alias{handle_tips_flow}
\title{Helper function to handle tips flow data.}
\usage{
handle_tips_flow(data)
}
\arguments{
\item{data}{A data frame containing tips flow data.}
}
\value{
A tibble with relevant information.
}
\description{
This function processes tips flow data and extracts relevant information.
}
