% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_rapidpro_key.R
\name{set_rapidpro_key}
\alias{set_rapidpro_key}
\title{Reset the key value in the environment}
\usage{
set_rapidpro_key(key)
}
\arguments{
\item{key}{string denoting the API Token associated with the website given in set_rapidpro_site()}
}
\value{

}
\description{
Reset the key value in the environment
}
