% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_rapidpro_site.R
\name{set_rapidpro_site}
\alias{set_rapidpro_site}
\title{Reset the website in the environment}
\usage{
set_rapidpro_site(site)
}
\arguments{
\item{site}{A string denoting the website associated with the token given in \code{set_rapidpro_key}.}
}
\value{
A string containing the website required to access RapidPro.
}
\description{
Reset the website in the environment
}
\examples{
#
}
