% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_rapidpro_uuid_names.R
\name{set_rapidpro_uuid_names}
\alias{set_rapidpro_uuid_names}
\title{Set the UUID names in the environment for the flow data}
\usage{
set_rapidpro_uuid_names(uuid_names = get_flow_names())
}
\arguments{
\item{uuid_names}{Default get_flow_names() (recommended). Data frame containing the flow names and UUIDs.}
}
\value{
Setting a data frame containing RapidPro flow data to the environment.
}
\description{
Set the UUID names to access data in RapidPro.
}
