% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unlist_rapidpro_variable.R
\name{unlist_rapidpro_variable}
\alias{unlist_rapidpro_variable}
\title{Unlist variables in a variable}
\usage{
unlist_rapidpro_variable(data, variable = "groups")
}
\arguments{
\item{data}{The data frame containing the data}

\item{variable}{The variable that contains the groups}
}
\value{
returns a data frame containing the unlisted data.
}
\description{
After importing data from RapidPro, unlist the variables that are stored in a list
}
