% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_archived_data.R
\name{update_archived_data}
\alias{update_archived_data}
\title{Update archived data}
\usage{
update_archived_data(
  curr_data,
  period = "none",
  date_to = NULL,
  format_date = "\%Y-\%m-\%d",
  tzone_date = "UTC"
)
}
\arguments{
\item{curr_data}{Archived data set from RapidPro (can be generated by \code{get_archived_data}).}

\item{period}{Default \code{"none"}. A character string giving the period of time. One of \code{"monthly"}, \code{"daily"}, or \code{"none"}.}

\item{date_to}{character string giving the date to filter the data to.}

\item{format_date}{from \verb{as.POSIX*} function: character string giving a date-time format as used by \code{strptime}.}

\item{tzone_date}{from \verb{as.POSIX*} function: time zone specification to be used for the conversion, if one is required. System-specific (see time zones), but "" is the current time zone, and "GMT" is UTC (Universal Time, Coordinated). Invalid values are most commonly treated as UTC, on some platforms with a warning.}
}
\value{
Updated archived data
}
\description{
Read in current archived data, and update it.
}
