% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_data_calculation.R
\name{flow_data_calculation}
\alias{flow_data_calculation}
\title{Calculate flow data.}
\usage{
flow_data_calculation(
  result_flow,
  flatten = FALSE,
  flow_type = "none",
  flow_handle_type = NULL,
  flow_handle_type_sub = "category",
  date_from = NULL,
  date_to = NULL,
  format_date = "\%Y-\%m-\%d",
  tzone_date = "UTC",
  created_on = FALSE
)
}
\arguments{
\item{result_flow}{Data frame containing flow data.}

\item{flatten}{Default \code{FALSE}. A boolean denoting whether the data should be flattened into a two-dimensional tabular structure.}

\item{flow_type}{Default \code{none}. Takes values \code{"praise"}, \code{"calm"}, \code{"check_in"}, \code{"tips"}, \code{"none"}. These are related to ParentText analysis.}

\item{flow_handle_type}{Default \code{NULL}. If \code{flow_type} is \code{other}, this is the value of the flow under \verb{$value} to get the response for.}

\item{date_from}{character string giving the date to filter the data from.}

\item{date_to}{character string giving the date to filter the data to.}

\item{format_date}{character string giving a date-time format as used by \code{strptime}.}

\item{tzone_date}{time zone specification to be used for the conversion, if one is required.}

\item{created_on}{Parameter for flow_data.}
}
\value{
Data frame displaying the flow data.
}
\description{
This function arranges flow data for different flow types.
}
