% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_flow_data.R
\name{get_flow_data}
\alias{get_flow_data}
\title{Get run (flow) data}
\usage{
get_flow_data(
  uuid_data = get_rapidpro_uuid_names(),
  flow_name,
  result,
  call_type = "runs.json?flow=",
  rapidpro_site = get_rapidpro_site(),
  token = get_rapidpro_key(),
  flatten = FALSE,
  checks = FALSE
)
}
\arguments{
\item{uuid_data}{A string containing UUID values. See set_rapidpro_uuid_names() to set this value.}

\item{flow_name}{A string containing flow names to call data from.}

\item{result}{TODO: result name? If so, we need to make it so you can change the result name.}

\item{call_type}{A string containing the call type.}

\item{rapidpro_site}{The rapidpro website to call the data from. See set_rapidpro_site() to amend the website.}

\item{token}{The token to call the data. See set_rapidpro_key() to amend the website.}

\item{flatten}{Default FALSE. A boolean denoting whether the data should be flattened into a two-dimensional tabular structure.}

\item{checks}{default FALSE. A boolean whether to check and update the data if it is not found.}
}
\value{

}
\description{
Get run (flow) data
}
