% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_flow_data.R
\name{get_flow_data}
\alias{get_flow_data}
\title{Get data from each run (flow)}
\usage{
get_flow_data(
  uuid_data = get_rapidpro_uuid_names(),
  flow_name,
  rapidpro_site = get_rapidpro_site(),
  token = get_rapidpro_key(),
  flatten = FALSE,
  checks = FALSE,
  return_all = FALSE,
  flow_type = "none",
  flow_handle_type = NULL,
  flow_handle_type_sub = "category",
  include_archived_data = FALSE,
  get_by = "gotit",
  data_from_archived = NULL,
  read_archived_data_from = NULL,
  created_on = FALSE,
  date_from = NULL,
  date_to = NULL,
  format_date = "\%Y-\%m-\%d",
  tzone_date = "UTC"
)
}
\arguments{
\item{uuid_data}{A string containing UUID values. See \code{set_rapidpro_uuid_names()} to set this value.}

\item{flow_name}{A string containing flow names to call data from.}

\item{rapidpro_site}{A string containing the rapidpro website to call the data from. See \code{set_rapidpro_site()} to amend the website.}

\item{token}{A string containing the token to call the data. See \code{set_rapidpro_key()} to amend the token.}

\item{flatten}{Default \code{FALSE}. A boolean denoting whether the data should be flattened into a two-dimensional tabular structure.}

\item{checks}{Default \code{FALSE}. A boolean whether to check and update the data if it is not found.}

\item{return_all}{Default \code{FALSE}. A boolean whether to return a specific part of the string, or all of the data.}

\item{flow_type}{Only if \code{return_all = FALSE}. Default \code{none}. Takes values \code{"praise"}, \code{"calm"}, \code{"check_in"}, \code{"tips"}, \code{"none"}, \code{"other"}. These are related to ParentText.}

\item{flow_handle_type}{Default \code{NULL}. If \code{flow_type} is \code{other}, this is the value of the flow under \verb{$value} to get the response for.}

\item{flow_handle_type_sub}{Default \code{NULL}. If \code{flow_type} is \code{other}, this is the group within \code{flow_handle_type} to handle.}

\item{include_archived_data}{Default \code{FALSE}. A boolean denoting whether to include archived data or not.}

\item{get_by}{A string denoting how to access the archived data (if \code{include_archived_data} is \code{TRUE}). Options are \code{"read_from_RDS"}, \code{"gotit"}. The first denotes that the data is read from an RDS file (given in \code{read_archived_data_from}), the latter that the data is stored and will be read in directly (see \code{data_from_archived}).}

\item{data_from_archived}{If \code{get_by = "gotit"}, this is the archived data.}

\item{read_archived_data_from}{If \code{get_by = "read_from_RDS"}, this is the archived data.}

\item{created_on}{Parameter for flow_data.}

\item{date_from}{character string giving the date to filter the data from.}

\item{date_to}{character string giving the date to filter the data to.}

\item{format_date}{from \verb{as.POSIX*} function: character string giving a date-time format as used by \code{strptime}.}

\item{tzone_date}{from \verb{as.POSIX*} function: time zone specification to be used for the conversion, if one is required. System-specific (see time zones), but "" is the current time zone, and "GMT" is UTC (Universal Time, Coordinated). Invalid values are most commonly treated as UTC, on some platforms with a warning.}
}
\value{
List separated by each flow_name provided. Each element in the list contains a data frame for each flow_name provided.
}
\description{
Call RapidPro data from each run or flow.
}
