% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_flow_names.R
\name{get_flow_names}
\alias{get_flow_names}
\title{Function to get a list of the UUID of flows.}
\usage{
get_flow_names(
  call_type = "flows.json",
  rapidpro_site = get_rapidpro_site(),
  token = get_rapidpro_key(),
  flatten = FALSE
)
}
\arguments{
\item{call_type}{A string containing the call type.}

\item{rapidpro_site}{The rapidpro website to call the data from. See set_rapidpro_site() to amend the website.}

\item{token}{The token to call the data. See set_rapidpro_key() to amend the website.}

\item{flatten}{Default FALSE. A boolean denoting whether the data should be flattened into a two-dimensional tabular structure.}
}
\value{

}
\description{
Function to get a list of the UUID of flows.
}
