% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rapidpro_site.R
\name{get_rapidpro_site}
\alias{get_rapidpro_site}
\title{Get the site value from the environment}
\usage{
get_rapidpro_site()
}
\value{
returns a string containing the website name set in the environment for the rapidPro data.
}
\description{
Get the site value from the environment
}
