% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_user_data.R
\name{get_user_data}
\alias{get_user_data}
\title{Get data for each user}
\usage{
get_user_data(
  call_type = "contacts.json",
  rapidpro_site = get_rapidpro_site(),
  token = get_rapidpro_key(),
  flatten = FALSE
)
}
\arguments{
\item{call_type}{A string containing the call type.}

\item{rapidpro_site}{A string containing the rapidpro website to call the data from. See \code{set_rapidpro_site()} to amend the website.}

\item{token}{A string containing the token to call the data. See \code{set_rapidpro_key()} to amend the token.}

\item{flatten}{Default \code{FALSE}. A boolean denoting whether the data should be flattened into a two-dimensional tabular structure.}
}
\value{
returns a data frame containing the contact data.
}
\description{
Call the contact data from RapidPro for each user
}
