% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_flows.R
\name{handle_type_flow}
\alias{handle_type_flow}
\title{Helper function to handle flow data of a specified type.}
\usage{
handle_type_flow(data, type = "completed", type_2 = "category")
}
\arguments{
\item{data}{A data frame containing flow data.}

\item{type}{The type of flow data to handle.}

\item{type_2}{The group within \code{type} to handle.}
}
\value{
A tibble with relevant information.
}
\description{
This function processes flow data of a specified type and extracts relevant information.
}
