% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httr_get_call.R
\name{httr_get_call}
\alias{httr_get_call}
\title{Calling data from RapidPro}
\usage{
httr_get_call(get_command, token = get_rapidpro_key())
}
\arguments{
\item{get_command}{A string containing the website to call the data from. Usually called within a function.}

\item{token}{A string containing the token to call the data. See set_rapidpro_key() to amend the token.}
}
\value{

}
\description{
Calling data from RapidPro
}
