% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_rapidpro_site.R
\name{set_rapidpro_site}
\alias{set_rapidpro_site}
\title{Reset the website in the environment}
\usage{
set_rapidpro_site(site)
}
\arguments{
\item{site}{A string denoting the website to call the API data from.}
}
\value{
A string containing the RapidPro website.
}
\description{
Set the website to access the data.
}
